requirements.txt
flask
flask-sqlalchemy
flask_wtf
wtforms.validators
wtforms
bleach
email_validator
flask_login
flask_bcrypt
flask_session